fix null

+++ phenex/phenotypes/computation_graph_phenotypes.py
@@ -115,7 +115,7 @@ class ComputationGraphPhenotype(Phenotype):
         # Add a null value column if it doesn't exist, for example in the case of a LogicPhenotype
         schema = joined_table.schema()
         if "VALUE" not in schema.names:
-            joined_table = joined_table.mutate(VALUE=ibis.null())
+            joined_table = joined_table.mutate(VALUE=ibis.null().cast("int32"))
 
         return joined_table
